Project info table logically split up
Each row in the table is now its own tag. Changes can now be made independently from each other (especially visual changes)
They are currently still all contained within the tags_info module, but this change enables them to be dynamically added.
