+++ bp/templates/bp/project_info_table_grade_info.html
@@ -1,0 +1,10 @@
+{% load fontawesome_5 %}
+
+{% if show_aggrade %}
+    <td>Bewertung:</td>
+    <td>
+        <b>{{ ag_points }} Punkte</b><br>
+        <br>
+        {{ ag_points_justification|linebreaks }}
+    </td>
+{% endif %}
